+++ CHANGES.rst
@@ -6,6 +6,7 @@ Changelog
 
 - switched from `imgaug3` to `imaug` for numpy 2 support
 - added `roi-images` filter for extracting sub-images based on object-detection bounding boxes
+- the `find-contours` filter now works on any type of image data
 
 
 0.0.9 (2025-04-03)